Don't index using a float

+++ glasbey.py
@@ -76,7 +76,7 @@ class Glasbey:
             self.palette = [self.colors[i, :] for i in self.palette]
         elif type(base_palette) == list and len(base_palette) > 0:
             self.palette = [(rgb[0] * 256 + rgb[1]) * 256 + rgb[2] for rgb in base_palette]
-            self.palette = [self.colors[np.round(i), :] for i in self.palette]
+            self.palette = [self.colors[int(np.round(i)), :] for i in self.palette]
         else:
             self.palette = [self.colors[-1, :]]  # white
